requirements.txt
aiogram==3.2.0
aiohttp==3.9.1
python-dotenv==1.0.0
websockets==12.0

